pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools >= 40.6.0",
    "wheel",
]

[project]
name = "fsec_extractor"
version = "1.0.0"
description = "Data extractor for F-SEC data."
readme = "README.md"
keywords = [
    "",
]
license = {file = "LICENSE"}
maintainers = []
authors = []
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
"numpy", "pandas", "matplotlib", "tkinter", "BaselineRemoval", "tensorflow", "keras.preprocessing"
]

[project.scripts]
"fsec_data_extractor" = "fsec_extractor.fsec_script:run"